requirements.txt
alembic==1.14.1
blinker==1.9.0
click==8.1.8
dnspython==2.7.0
Flask==2.1.3
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.8
MarkupSafe==3.0.2
PyJWT==2.10.1
python-dotenv==0.19.0
SQLAlchemy==1.4.23
typing_extensions==4.12.2
Werkzeug==2.0.1
WTForms==3.2.1
zipp==3.21.0
